fix(user): prevent user from deleting their own account

+++ server/user/controllers/user.controller.ts
@@ -109,7 +109,10 @@ export default class UserController implements CrudController<User> {
   @UseGuards(AuthGuard('jwt'))
   @UseInterceptors(CrudRequestInterceptor)
   @Delete('/:id')
-  async delete(@Param('id') id: string) : Promise<void> {
+  async delete(@Param('id') id: string, @Request() { user }) : Promise<void> {
+    if (user.id === Number(id)) {
+      throw new HttpException('CANNOT_DELETE_SELF', 403);
+    }
     await this.service.deleteUser(id);
   }
